Move biip.{gs1 => symbology}.GS1Symbology
Breaking change
